# improve some IntSet operations