* IntSet has a new invariant: the length of the underlying BitArray
  is always a multiple of 64
* the size is set to the minimum possible: the underlying routines
  are in charge of allocating enough (like for other containers);
  this can help to obtain the last element faster (which is useful
  e.g in `rand(::IntSet)`)
* _matched_map avoids extending a BitArray when unecessary
  (but doesn't work anymore for all 2^4 kinds of binary predicate,
  though it could be handled if needed in the future)

A case where this could be slower is for example in:
```
s = IntSet()
for i=1:1000
    push!(s, i)
end
```
But I could not really/reliably measure a difference (disclaimer: `BenchmarkTools` is broken recently on my installation, so I use a hand-made "equivalent"), and `sizehint!` can help here. The improvements for some operations can be huge, e.g.  `symdiff!(a, b)` where `a=IntSet([2^20]), b=IntSet([1])` decreases from 45.07µs to 10ns. Also, something like `rand(s)` where `s=IntSet([1, 2, 3])` decreases from about 980ns to 310ns.