@nanosoldier 2017-08-05
[Your benchmark job](https://github.com/JuliaLang/julia/pull/23138#issuecomment-320467990) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/3dbb08727df11be632005c24020215ff4473245f/e181e2c_vs_83007fb/report.md). cc @ararslan