@tkelman 2017-08-06
stockcorr and array indexing regressions have a chance of being real, should keep an eye on the nightly results