@fredrikekre 2017-08-06
Given https://github.com/JuliaLang/julia/commit/ae1719872092b594228a9bfeedfc3fa8e91c9529#commitcomment-23501641 it might be worth looking into the array indexing regression?