@rfourquet 2017-08-06
Looking at the very few places in base where `IntSet` are used, I have a hard time understanding how this change could affect these 2 tests... Also, I cannot reproduce locally, i.e. the regressions on those two tests, "stockcorr" and "sumvector_view", stay within 5% (for some reason, the `@benchmark` macro fails in my REPL, but running BaseBenchmarks is ok :-/ )