@tkelman 2017-08-06
doesn't inference use them in some key places? otherwise they'd likely be moved out to DataStructures