@rfourquet 2017-08-06
Inference uses them, but I thought it wouldn't influence the run time performance and the benchmarks.